Adding the ability to divide the wedge with vertical lines for the purposes of the subduction zone benchmark diagnostics.
Also tidying up the default params and making the benchmark geometry respect them so that the benchmark can be standardized.

The data example has also had these modifications added just to check they work on arbitrary geometries.

Finally adding a feature to the geometry plotting that allows sids and rids to be visualized (not very well).