Finally managed to silence gmsh!
Stop talking about how noisy gmsh is.

Fix the all_sz.json data file by adding Ac and As (they are subtracted in the code in a way that they are not in TF)

Update names of the steady state suite.  Run them at a higher resolution (will see how slow that is).

Use find in clear_notebooks and update_python.

Add a file to keep track of typos we find.