Fixing an error with kc and rhoc in sz_problem.
Adding a warning if kc, rhoc, H1 or H2 are modified and the case is oceanic as they're not used.

Updating all the global suite to compare to the vtus from 2023.

Adding pyvista diffing routines and classes to utils.

Adding the 2023 vtus using git lfs (all vtus are automatically tracked using git lfs).

Adding git lfs to the Dockerfile (needs to be rebuilt still).

Specifying the website as english.